Split main data content block and separate metadata block
