+++ decidim-core/app/cells/decidim/content_blocks/participatory_space_main_data/content.erb
@@ -3,9 +3,3 @@
 </div>
 
 <%= nav_links %>
-
-<div class="<%= prefixed_class("metadata") %>">
-  <%= render :metadata %>
-
-  <%= render :recent_users %>
-</div>
