+++ decidim-core/app/cells/decidim/content_blocks/participatory_space_main_data_cell.rb
@@ -20,10 +20,6 @@ module Decidim
       def nav_items
         []
       end
-
-      def metadata_items
-        []
-      end
     end
   end
 end
