+++ decidim-core/app/cells/decidim/content_blocks/participatory_space_metadata/content.erb
@@ -1,0 +1,5 @@
+<div class="<%= prefixed_class("metadata") %>">
+  <%= render :metadata %>
+
+  <%= render :recent_users %>
+</div>
