+++ decidim-core/app/cells/decidim/content_blocks/participatory_space_metadata/metadata.erb
