+++ decidim-core/app/cells/decidim/content_blocks/participatory_space_metadata/recent_users.erb
