+++ decidim-core/app/cells/decidim/content_blocks/participatory_space_metadata_cell.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Decidim
+  module ContentBlocks
+    class ParticipatorySpaceMetadataCell < BaseCell
+      private
+
+      def metadata_items
+        []
+      end
+    end
+  end
+end
