+++ decidim-participatory_processes/app/cells/decidim/participatory_processes/content_blocks/main_data_cell.rb
@@ -8,10 +8,8 @@ module Decidim
         include ParticipatoryProcessHelper
         include Decidim::ComponentPathHelper
         include ActiveLinkTo
-        include Decidim::SanitizeHelper
-        include Decidim::ModalHelper
 
-        delegate :short_description, :steps, :active_step, :start_date, :end_date, :participatory_process_group, to: :resource
+        delegate :short_description, to: :resource
 
         private
 
@@ -29,53 +27,6 @@ module Decidim
 
         def nav_items
           process_nav_items(resource)
-        end
-
-        def metadata_items
-          [step_metadata_item, dates_metadata_item, group_item].compact
-        end
-
-        def classes_prefix
-          "process"
-        end
-
-        def active_step_name
-          translated_attribute active_step.title
-        end
-
-        def step_metadata_item
-          {
-            title: t("active_step", scope: "layouts.decidim.participatory_processes.participatory_process"),
-            icon: "direction-line",
-            partial: "active_step"
-          }
-        end
-
-        def dates_metadata_item
-          {
-            title: [
-              t("start_date", scope: "activemodel.attributes.participatory_process_step"),
-              t("end_date", scope: "activemodel.attributes.participatory_process_step")
-            ].join(" / "),
-            icon: "calendar-todo-line",
-            text: [
-              start_date.present? ? l(start_date, format: :decidim_short_with_month_name_short) : "?",
-              end_date.present? ? l(end_date, format: :decidim_short_with_month_name_short) : "?"
-            ].join(" / ")
-          }
-        end
-
-        def group_item
-          return if participatory_process_group.blank?
-
-          {
-            title: t("belongs_to_group", scope: "decidim.participatory_processes.show"),
-            icon: "archive-line",
-            text: link_to(
-              translated_attribute(participatory_process_group.title),
-              decidim_participatory_processes.participatory_process_group_path(participatory_process_group)
-            )
-          }
         end
 
         def extra_classes
