+++ decidim-participatory_processes/app/cells/decidim/participatory_processes/content_blocks/metadata/active_step.erb
