+++ decidim-participatory_processes/app/cells/decidim/participatory_processes/content_blocks/metadata_cell.rb
@@ -1,0 +1,64 @@
+# frozen_string_literal: true
+
+module Decidim
+  module ParticipatoryProcesses
+    module ContentBlocks
+      class MetadataCell < Decidim::ContentBlocks::ParticipatorySpaceMetadataCell
+        include ParticipatorySpaceContentBlocksHelper
+        include ParticipatoryProcessHelper
+        include Decidim::ModalHelper
+
+        delegate :steps, :active_step, :start_date, :end_date, :participatory_process_group, to: :resource
+
+        private
+
+        def metadata_items
+          [step_metadata_item, dates_metadata_item, group_item].compact
+        end
+
+        def active_step_name
+          translated_attribute active_step.title
+        end
+
+        def step_metadata_item
+          {
+            title: t("active_step", scope: "layouts.decidim.participatory_processes.participatory_process"),
+            icon: "direction-line",
+            partial: "active_step"
+          }
+        end
+
+        def dates_metadata_item
+          {
+            title: [
+              t("start_date", scope: "activemodel.attributes.participatory_process_step"),
+              t("end_date", scope: "activemodel.attributes.participatory_process_step")
+            ].join(" / "),
+            icon: "calendar-todo-line",
+            text: [
+              start_date.present? ? l(start_date, format: :decidim_short_with_month_name_short) : "?",
+              end_date.present? ? l(end_date, format: :decidim_short_with_month_name_short) : "?"
+            ].join(" / ")
+          }
+        end
+
+        def group_item
+          return if participatory_process_group.blank?
+
+          {
+            title: t("belongs_to_group", scope: "decidim.participatory_processes.show"),
+            icon: "archive-line",
+            text: link_to(
+              translated_attribute(participatory_process_group.title),
+              decidim_participatory_processes.participatory_process_group_path(participatory_process_group)
+            )
+          }
+        end
+
+        def extra_classes
+          prefixed_class("content-block")
+        end
+      end
+    end
+  end
+end
