+++ decidim-participatory_processes/app/helpers/decidim/participatory_processes/participatory_space_content_blocks_helper.rb
@@ -12,6 +12,10 @@ module Decidim
       def decidim_participatory_processes
         Decidim::ParticipatoryProcesses::Engine.routes.url_helpers
       end
+
+      def classes_prefix
+        "process"
+      end
     end
   end
 end
