+++ decidim-participatory_processes/lib/decidim/participatory_processes/engine.rb
@@ -114,6 +114,12 @@ module Decidim
           content_block.default!
         end
 
+        Decidim.content_blocks.register(:participatory_process_homepage, :metadata) do |content_block|
+          content_block.cell = "decidim/participatory_processes/content_blocks/metadata"
+          content_block.public_name_key = "decidim.content_blocks.metadata.name"
+          content_block.default!
+        end
+
         Decidim.content_blocks.register(:participatory_process_homepage, :last_activity) do |content_block|
           content_block.cell = "decidim/participatory_processes/content_blocks/last_activity"
           content_block.public_name_key = "decidim.content_blocks.last_activity.name"
